Add logo Update search bar
