+++ package.json
@@ -12,13 +12,13 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.5",
+    "@vitejs/plugin-react": "^4.0.3",
     "axios": "^1.4.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.2",
-    "@vitejs/plugin-react": "^4.0.3",
     "vite": "^4.4.5"
   },
   "devDependencies": {
