+++ src/components/CoinCard.jsx
@@ -1,21 +1,16 @@
-/* eslint-disable react/prop-types */
 import { Link } from 'react-router-dom';
-
-// import PropTypes from 'prop-types';
+import PropTypes from 'prop-types';
 
 const CoinCard = ({ coin, index }) => {
-  const isEvenIndex = index % 2 === 0;
-  const cardClass = isEvenIndex ? 'even' : 'odd';
-  const isBgColor = isEvenIndex ? 'odd' : 'even ';
-
+  const isIndex = (index === 0 || index === 3 || index === 4 || index === 7 || index === 8);
+  const isColor = isIndex ? 'odd' : 'even';
   return (
-    <div className={`col-4 col-sm-4 col-md-4 col-lg-4 zero-pad ${cardClass}`}>
-      <div className={`card border-0 rounded-0 m-0 ${cardClass}`}>
+    <div className={`col-6 col-sm-6 col-md-6 col-lg-6 zero-pad ${isColor}`}>
+      <div className={`card border-0 rounded-0 m-0 ${isColor}`}>
         <div className="align-items-sm-center card-body d-flex flex-column">
           <div className="align-self-center">
-
             <Link to={`/coins/${coin.id}`}>
-              <img className={`iconStyle p-1 rounded-5 ${isBgColor}`} src={coin.icon} alt="" />
+              <img className={`iconStyle p-1 rounded-5 ${isColor}`} src={coin.icon} alt="" />
             </Link>
           </div>
           <div className="">
@@ -24,9 +19,7 @@ const CoinCard = ({ coin, index }) => {
           <div className="align-self-center">
             <div className="fs-6 text-center text-white">
               Price:
-              {' '}
               <br />
-              {' '}
               {coin.price.toFixed(2)}
             </div>
           </div>
@@ -34,6 +27,16 @@ const CoinCard = ({ coin, index }) => {
       </div>
     </div>
   );
+};
+
+CoinCard.propTypes = {
+  coin: PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    icon: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+    price: PropTypes.number.isRequired,
+  }).isRequired,
+  index: PropTypes.number.isRequired,
 };
 
 export default CoinCard;
