+++ src/components/CryptoSearch.jsx
@@ -2,71 +2,51 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useEffect } from 'react';
 import { fetchCoin, setGlobalSearch } from '../redux/slices/coinSlice';
 import './styles/cryptoSearch.css';
+import cryptos from '../cryptos.png';
 
 const CryptoSearch = () => {
+  const coinData = useSelector((state) => state.coinSlice);
   const dispatch = useDispatch();
-  const coinData = useSelector((state) => state.coinSlice.coin);
-  const isLoading = useSelector((state) => state.coinSlice.isLoading);
-  const error = useSelector((state) => state.coinSlice.error);
 
-  const handleFormSubmit = (e) => {
-    e.preventDefault();
-    const searchValue = e.target.elements.searchInput.value.trim();
-    dispatch(setGlobalSearch(searchValue));
-    console.log(searchValue);
+  const handleInputChange = (e) => {
+    const global = e.target.value.trim();
+    dispatch(setGlobalSearch(global));
+    console.log(global);
   };
+
+  const handleKeyPress = (e) => {
+    if (e.key === 'Enter') {
+      e.preventDefault();
+    }
+  };
+
   useEffect(() => {
-    if (!coinData.length) {
+    if (coinData.length === 0) {
       dispatch(fetchCoin());
     }
-  }, [dispatch, coinData.length]);
-
-  if (isLoading) {
-    return (
-      <div className="container">
-        <div className="row">
-          <div className="col">
-            <h2 data-testid="h2test" className="align-items-center d-flex justify-content-center text-light vh-100">Loading...</h2>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  if (error) {
-    return (
-      <div className="container">
-        <div className="row">
-          <div className="col">
-            <div className="align-items-center d-flex justify-content-center text-danger vh-100 fs-2">
-              {error}
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  }
+  }, [dispatch, coinData]);
 
   return (
-    <div className="background py-2">
-      <form onSubmit={handleFormSubmit} className="container">
-        <div className="row align-items-center">
-          <div className="col-8 col-sm-8">
-            <input
-              type="text"
-              name="searchInput"
-              className="form-control"
-              placeholder="Bitcoin, Ethereum, etc."
-            />
+    <>
+      <div className="d-flex justify-content-center img-container">
+        <img className="img-fluid w-100 img-max-width" src={cryptos} alt="" />
+      </div>
+      <div className="background py-2">
+        <form className="container">
+          <div className="row align-items-center">
+            <div className="col-12 col-sm-12">
+              <input
+                onChange={handleInputChange}
+                onKeyPress={handleKeyPress}
+                type="text"
+                className="form-control"
+                placeholder="Bitcoin, Ethereum, etc."
+              />
+            </div>
           </div>
-          <div className="col-2 col-sm-4">
-            <button type="submit" className="btn btn-primary">
-              Search
-            </button>
-          </div>
-        </div>
-      </form>
-    </div>
+        </form>
+      </div>
+    </>
   );
 };
